requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-generator==1.10
async-timeout==5.0.1
atomicwrites==1.4.1
attrs==24.2.0
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
Django==5.1.4
django-debug-toolbar==4.4.6
django-elasticsearch-dsl==8.0
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
docker==7.1.0
elastic-transport==8.15.1
elasticsearch==8.17.0
elasticsearch-dsl==8.17.0
factory_boy==3.3.1
Faker==33.1.0
flower==2.0.1
h11==0.14.0
humanize==4.11.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
py==1.11.0
pycparser==2.22
pyOpenSSL==24.3.0
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.4
pytest-base-url==2.1.0
pytest-django==4.9.0
pytest-factoryboy==2.7.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-selenium==4.1.0
pytest-variables==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.2
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
